refactor : @types에서 types로 폴더 이동
