+++ types/index.ts
@@ -1,0 +1,1 @@
+export type PageType = 'grouping' | 'meeting' | 'profile';
